link-land Link now shows up on time automatically.
