+++ js/LinkLand/LinkLand.jsx
@@ -11,25 +11,14 @@ class LinkLand extends React.PureComponent {
         this.saveLink = this.saveLink.bind(this);
         this.retrieveLink = this.retrieveLink.bind(this);
         this.clearLinks = this.clearLinks.bind(this);
-        this.toggleHidden = this.toggleHidden.bind(this);
 
         let links = localStorage.getItem("links");
         links = links ? JSON.parse(links) : [];
 
         this.state = {
             links: links,
-            isHidden: true
+            currentLink: ''
         }
-    }
-
-    toggleHidden () {
-        var start = DateTime.fromObject({hour: 10, minutes: 15});
-        var end = DateTime.fromObject({hour: 10, minutes: 20});
-        var i = Interval.fromDateTimes(start, end);
-        if (i.contains(DateTime.local()) && this.state.isHidden)
-            this.setState({
-              isHidden: !this.state.isHidden
-            })
     }
 
     saveLink(event) {
@@ -51,10 +40,9 @@ class LinkLand extends React.PureComponent {
         const links = [...this.state.links];
         if (links.length) {
             const linkToRetrieve = links.shift();
-            const input = document.getElementById("linkLandInput");
-            input.value = linkToRetrieve;
             this.setState({
-                links: links
+                links: links,
+                currentLink: linkToRetrieve
             });
             setLinks(links);
         }
@@ -69,22 +57,11 @@ class LinkLand extends React.PureComponent {
         setLinks([]);
     };
 
-    componentDidUpdate() {
-        this.toggleHidden();
-    }
-
-    componentDidMount() {
-        this.toggleHidden();
-    }
-
     render() {
-        const {
-
-        } = this.props;
         return (
             <div>
-                <ViewLink isHidden={this.state.isHidden} link={'https://www.youtube.com/watch?v=uCLEq9V0-Yk'}/>
-                <LinkInput isHidden={this.state.isHidden} saveLink={this.saveLink} retrieveLink={this.retrieveLink} clearLinks={this.clearLinks}/>
+                <ViewLink  link={this.state.currentLink} retrieveLink={this.retrieveLink}/>
+                <LinkInput saveLink={this.saveLink} retrieveLink={this.retrieveLink} clearLinks={this.clearLinks}/>
             </div>
         );
     }
